requirements.txt
absl-py==1.2.0
aniso8601==9.0.1
asgiref==3.4.1
astunparse==1.6.3
Brotli==1.0.9
cachetools==5.2.0
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cmake==3.22.4
colorama==0.4.4
ComboFinder==0.1.1
cryptography==37.0.4
dash==2.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.5
dj-database-url==1.0.0
Django==3.2.6
django-crispy-forms==1.11.1
django-environ==0.8.1
django-gunicorn==0.1.1
django-heroku==0.3.1
django-model-utils==3.1.2
et-xmlfile==1.1.0
face-recognition-models==0.3.0
filelock==3.7.1
Flask==2.1.2
Flask-Compress==1.11
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
gast==0.4.0
gcloud==0.18.3
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.48.0
gunicorn==20.1.0
h5py==3.7.0
httplib2==0.20.4
idna==2.10
importlib-metadata==4.11.3
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3.1
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
lxml==4.8.0
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
multitasking==0.0.10
mysql==0.0.3
mysql-connector-python==8.0.29
mysqlclient==2.1.0
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.0
odfpy==1.4.1
opencv-python==4.5.5.64
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandas-datareader==0.10.0
Pillow==8.3.2
pipenv==2022.7.24
platformdirs==2.5.2
plotly==5.7.0
prettytable==3.3.0
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pyparsing==3.0.9
Pyrebase4==4.5.0
pytesseract==0.1.9
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jwt==3.3.3
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
SQLAlchemy==1.4.39
sqlparse==0.4.1
tablib==3.2.1
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing_extensions==4.1.1
urllib3==1.26.3
virtualenv==20.15.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.1.2
whitenoise==6.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
yfinance==0.1.70
zipp==3.8.0
